Moved input buffer from UI and gave PlayerEntity an InputBuffer
